# docs/swagger.yaml
definitions:
  ds.Meals:
    properties:
      id:
        type: integer
      image_url:
        type: string
      meal_brand:
        type: string
      meal_detail:
        type: string
      meal_info:
        type: string
      meal_weight:
        type: string
      status:
        type: boolean
    type: object
  ds.MilkRequests:
    properties:
      address:
        type: string
      creator:
        $ref: '#/definitions/ds.Users'
      creator_id:
        type: integer
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      delivery_date:
        type: string
      id:
        type: integer
      moderator:
        $ref: '#/definitions/ds.Users'
      moderator_id:
        type: integer
      recipient_name:
        type: string
      recipient_surname:
        type: string
      status:
        type: integer
    type: object
  ds.Users:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
...
  schemas.FinishMilkRequestRequest:
    properties:
      id:
        type: string
    type: object
...
  schemas.GetAllMilkRequestsWithParamsResponse:
    properties:
      milkRequests:
        items:
          $ref: '#/definitions/ds.MilkRequests'
        type: array
    type: object
...
  schemas.GetMilkRequestResponse:
    properties:
      count:
        type: integer
      milk_request_meals:
        items:
          $ref: '#/definitions/ds.Meals'
        type: array
      milk_requests:
        $ref: '#/definitions/ds.MilkRequests'
    type: object
...
  schemas.ResponseMessage:
    type: object
...
info:
  contact: {}
paths:
...
  /api/milk_request/{ID}:
    delete:
      consumes:
      - application/json
      description: Удаляет заявку на молочную кухню по её ID
      parameters:
      - description: ID заявки
        in: path
        name: ID
        required: true
        type: integer
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: MilkRequest was deleted"  // Сообщение об успешном удалении
          schema:
            type: string
        "400":
          description: Неверный ID"  // Ошибка при неверном ID
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Ошибка сервера"  // Ошибка сервера при удалении заявки
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth  // Требуется аутентификация: []
      summary: Удалить заявку на молочную кухню
      tags:
      - milk_requests
    get:
      consumes:
      - application/json
      description: Получить данные заявки на молочную кухню по её ID, включая список
        связанных блюд
      parameters:
      - description: ID заявки
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetMilkRequestResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Получить данные заявки на молочную кухню
      tags:
      - milk_requests
  /api/milk_request/finish/{ID}:
    put:
      consumes:
      - application/json
      description: Завершаем заявку на молочную кухню по переданному ID и параметрам
        в теле запроса (статус и дата доставки)
      parameters:
      - description: ID заявки
        in: path
        name: ID
        required: true
        type: integer
      - description: Данные для завершения заявки
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/schemas.FinishMilkRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Milk Request was Finished
          schema:
            type: string
        "400":
          description: Ошибка в параметрах запроса
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Завершить заявку на молочную кухню
      tags:
      - milk_requests
  /api/milk_request/form/{ID}:
    put:
      consumes:
      - application/json
      description: Формирует заявку на молочную кухню по переданному ID и параметрам
        запроса
      parameters:
      - description: ID заявки
        in: path
        name: ID
        required: true
        type: integer
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Milk Request was Formed
          schema:
            type: string
        "400":
          description: Ошибка в параметрах запроса
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Создать заявку на молочную кухню
      tags:
      - milk_requests
  /api/milk_requests:
    get:
      consumes:
      - application/json
      description: Получить список запросов на молоко с возможностью фильтрации по
        статусу и пользователю
      parameters:
      - description: Статус заявки
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetAllMilkRequestsWithParamsResponse'
        "400":
          description: Неверный запрос, отсутствует параметр status или он некорректен
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/schemas.ResponseMessage'
      security:
      - BearerAuth: []
      summary: Получить все заявки на молочную кухню с параметрами
      tags:
      - milk_requests
...
swagger: "2.0"